Temporarily disable signature check when targets metadata exists
This is necessary because in the current state model, we rely on the
parent state to indicate valid signing keys for a role. However, when
policies are changed, existing metadata for some states may become
invalid. First, the signature check is guaranteed to fail because of
new expectations. Second, if those refs are not updated immediately, we
apply invalid policies to those changes rather than trace back to what
was valid when the state was originally updated. Decoupling policies
from the state log will fix this.

Ticket: https://github.com/adityasaky/gittuf/issues/13

Signed-off-by: Aditya Sirish <aditya@saky.in>