Define RSL entry types
This change defines a regular RSL entry (with a change in format from
what was defined before) and an annotation type that can be used to
refer to one or more previous RSL entries. The annotation type can
include human friendly messages and can indicate if the entries referred
to must be skipped by gittuf clients.

This commit also moves the rsl package into the internal namespace.

Signed-off-by: Aditya Sirish <aditya@saky.in>