Add plumbing for policy
The tuf package includes models of TUF metadata with some minor
modifications. Largely, they are based on what's defined in go-tuf.

The signerverifier package provides some plumbing for use with DSSE to
sign the TUF metadata. Within it, the ed25519 package creates a DSSE
compatibler signer and verifier. The dsse package provides opinionated
APIs to create and verify signed metadata.

Signed-off-by: Aditya Sirish <aditya@saky.in>