common: Remove package entirely
The package offered two functions: GetRepositoryHandler and
CreateTestRepository. With a series of other changes,
GetRepositoryHandler is only needed once in the repository package.
Other uses in the cmd dev subcommand will be dropped in future.
Similarly, all tests have moved to either in-memory repositories or to
using the gittuf repository itself (one instance). Therefore, this
package can now be removed.

Signed-off-by: Aditya Sirish <aditya@saky.in>