cmd: Refactor entirely
This is a full fledged refactor of the cmd package based on this comment:
https://github.com/adityasaky/gittuf/pull/41#discussion_r1198995982.

First, the dev subcommand is retired as it was necessary to test certain
RSL functions which have now been built. The subcommand also printed out
the git config, again something we no longer need. To that end,
GetConfig() in gitinterface has been made private.

Second, the trust and policy subcommands have been reimplemented using
the recommended structure with specific packages that construct them
with their options.

Finally, the cmd package has been moved to the internal namespace.

Signed-off-by: Aditya Sirish <aditya@saky.in>