policy: Add basic verification functions
This commit introduces basic verification workflows for Git refs.
Specifically, a Git ref can be verified against either the latest policy
or from the very start of the RSL. As part of this, relative
verification is also introduced from one RSL entry to the next.

Note: also adds basic Match functionality for delegation patterns in the tuf
package.

Signed-off-by: Aditya Sirish <aditya@saky.in>