*: Add RSL APIs for fetching range of entries
This commit adds two APIs to the RSL package to fetch entries (and
relevant annotations) between a specified range. The policy package is
updated to use this functionality when verifying a range of entries
rather than walking the RSL there. This API enables future extensibility
of the policy package to handle annotations during verification.

Signed-off-by: Aditya Sirish <aditya@saky.in>