*: Vendor go-git with fix for fetch
This commit vendors go-git until a patch submitted upstream
(https://github.com/go-git/go-git/pull/875) is merged and is available
to us via a release. The patch is applied to this copy of go-git.

Steps taken to vendor:
1. go mod vendor -> vendors all dependencies
2. Copy go-git to the internal namespace
3. Run go mod init with go-git's module name and go mod tidy in the
   internal copy of go-git
4. Add FORK.md file
5. Add replace directive to gittuf's go.mod followed by go mod tidy
6. Apply patch to internal go-git
7. Remove vendor directory

Having patched go-git, this commit also enables two previously disabled
tests that relied on the corrected behavior in go-git.

Signed-off-by: Aditya Sirish <aditya@saky.in>