policy: Apply latest policy for file changes
Previously, we were trying to identify when a commit was first
introduced into the RSL to use the file policies applicable at the time.
This makes sense for handling cases such as a commit creator leaving an
organization prior to it being merged into a branch. However, this goes
against expectations: when a policy is changed, the expectation is that
all future changes use that policy.

We can solve the earlier issue with policy rotations invalidating prior
changes being merged now using attestation signatures. Now, the commit
creator's signature does not matter as long as it is approved via an
attestation by whoever is trusted in the current policy.

Signed-off-by: Aditya Sirish <aditya@saky.in>