policy: Handle policy-staging edge cases
This commit changes a number of things. First, it updates LoadState to
handle cases where there are no policy entries and the requested entry
is for the staging ref. This is going to be the case at the start of
gittuf use, when setting up the first policy. Second, it updates
LoadState to verify policy states that exist and then return the staging
ref if that's what is requested. What this means is that even when
trying to load a staging policy state, we verify that the main policies
are in a good state. Third, it removes state.Verify calls in Commit()
and the loadStateForEntry helper. Verify will fail quite often when
Commit() is called simply because we don't have sufficient signatures,
we're committing to staging now after all. loadStateForEntry is, for
similar reasons, not verifying anymore because its purpose is to load
the state into memory from the git object store. To compensate,
LoadState now calls Verify for policy states and not for staging states.
Finally, Apply() now calls Verify explicitly because this is operating
on a staging state and we want to know that the state is internally
consistent before applying to the policy ref.

This commit, while marked as updating the policy package, also updates
some calls to the policy package in repository because the context
parameter is no longer needed for Commit as it is no longer invoking
state.Verify.

Signed-off-by: Aditya Sirish <aditya@saky.in>