*: Remove TUF fields that don't apply to gittuf
The original root and targets metadata schemas were taken directly from
the TUF specification. As such, they include fields that we don't use in
gittuf.

Removed Fields:

1. SpecVersion: this field identifies the version of the TUF
   specification the metadata conforms to. We don't conform to the TUF
   specification.

2. ConsistentSnapshot: this boolean field indicates if the repository
   uses consistent snapshots as defined in the TUF specification. We
   don't use this (and in fact we don't use snapshot metadata at all as
   the RSL effectively serves the snapshot role).

3. Version: this field is an incrementing integer that identifies the
   version of the metadata for the specific role. In TUF, this is used
   to ensure we have a consistent set of metadata via the snapshot role.
   As before, the RSL gives us this property naturally.

Note that we still don't use the Expires field in gittuf verification.
We need to explore what it means for policy metadata to expire in a Git
repository and whether the expiry check (for the latest policy metadata)
applies the same way it does in TUF.

Signed-off-by: Aditya Sirish <aditya@saky.in>