*: Sign over name in targets files [DO NOT MERGE]
This can mitigate some confusion attacks where an incorrect targets
metadata file is returned during policy manipulations / delegation graph
walks.

NOTE: this needs some notion of gittufVersion in the metadata when we
introduce the check to see if the name in the metadata is as expected.
Metadata generated prior to this feature's release will have an empty
string that we should not error on.

Signed-off-by: Aditya Sirish <aditya@saky.in>