*: Update code review approval layer to be forge agnostic and clarify docs
The code review approval layer is now agnostic to the code review system used.
The only system supported at the moment is GitHub's pull request approvals, but
we should be able to extend this to support GitLab, Gerrit, etc.

In addition, the GitHub app key add endpoint docs have been updated to clarify
what the purpose of the key is.

Signed-off-by: Aditya Sirish A Yelgundhalli <ayelgundhall@bloomberg.net>