transport: Simplify flow and switch to APIs
This commit reorganizes the remote helpers to have an easier to follow
flow of execution. It also moves the transport-related documentation to
the corresponding package for better discoverability.

Additionally, some invocations of exec.Command have been replaced by
APIs.

Finally, the logic for when to update local gittuf refs has been updated
to first check if the local is ahead of the remote. This uses the newly
introduced reconciliation function in repository.

Signed-off-by: Aditya Sirish <aditya@saky.in>