rsl: Add incrementing number to RSL entry
This commit adds a uint64 Number field to RSL entries. This allows us to
determine the ordering of distinct RSL entries, enabling optimizations
such as knowing which policy to use for an entry without always walking
the RSL.

The implementation maintains backwards compatibility, so the numbering
can start from whenever this is enabled. This is made easy by the fact
that we can guarantee gittuf is updated everywhere it is currently used.

Signed-off-by: Aditya Sirish <aditya@saky.in>