gitinterface: Fix merge trees and support restore
This change ensures merge tree can be computed when the base commit is
zero (falling back to a fast forward merge).

Additionally, it adds support for restoring "staged changes" so we can
have a clean worktree for testing with older Git versions where we
actually have to use git checkout to compute merge trees.

Signed-off-by: Aditya Sirish <aditya@saky.in>