policy: Fix attestation load for tag verification
Currently, attestations are loaded assuming the approval target is a Git tree.
Entries for tags don't have trees. Separately, we can add support for approval
attestations for tags that know not to compute merge trees when approving a
tag, instead using the commit ID the tag is expected to point to.

Signed-off-by: Aditya Sirish A Yelgundhalli <ayelgundhall@bloomberg.net>