policy: Streamline verification helper
We have two very similar helpers: verifyGitObjectAndAttestations and
verifyGitObjectAndAttestationsUsingVerifiers. The first one calls the
second one. However, there were numerous cases of direct invocation of
the second one. This commit makes the first helper the one that's used
consistently. This ensures that finding which verifiers to use and so on
are consistent everywhere.

Signed-off-by: Aditya Sirish <aditya@saky.in>