policy: Fix break scope in VerifyRelativeForRef
We had a break statement to exit the loop looking for fix entries when a
policy violation was found. When we introduced propagation entries, we
added a switch-case block within the loop but did not update the break
statement to apply to the loop. This put the client in an invalid state:
a fix for the policy violation is found but it continues in recovery
mode eternally.

Signed-off-by: Aditya Sirish <aditya@saky.in>